refactor: extract image preloading logic into custom hook
